server/auth/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
facebook-sdk==3.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
